FinalTask/requirements.txt
pyTelegramBotAPI
Flask
python-dotenv
matplotlib
